First iteration of neovim configuration
LSP configuration for terraform and python. Python is using pyright,
for which Microsoft won't add snippet support in the open source
package. What this means is that the autocompletion support mimics
something you found in editors around 1970.

Treesitter configuration seems to work just fine.

Separated configuration to multiple files. With plugins this should
enable me to easily remove everything related to a plugin.