git: Moving away from worktrees
This requires further tooling. You don't have access to remotes, tooling
needs to change $PWD on branch changes. This needs support in tmux side
to switch PWD on all panes and restart services. Not having access to
remotes is not doable, thus at least bare repos are out the window.