dunst: fix user bin paths related to dunst
Discord and Spotify view notifications as the most important thing in
the world, they will block the whole UI and make the application
unusable if a notification cannot be delivered. Because, that is the
reason people use those applications, content and features they produce
is meaningless, only the notifications matter, praise the emperor
libnotify.

Anyways. Fix dunst startup.