nixos: Enable i3 and X11
Implements the same flow I had on arch linux where the system boots to
vtty and then I startx to i3 if needed. Installs rest of the missing
dotfiles as well, mostly just software I need in i3.

Note that this does not yet boot properly as I don't have drivers for
NVidia.