shell: show nix shell state in prompt
I run nix shells via `nix develop -c zsh`. It's easy to end up neck deep
in these. When I was testing this change I eventually noticed I was like
8 shells deep in nested nix shells.

While there's not much harm done if this happens it's still quite
annoying. And you are in a shell with extra functionality than in the
regular shell.