vm: add iso build for work environment
This builds what I use in the desktop minus some of the personal GUI
apps like discord and steam.

The idea is to use the VM for development and then in the host system
I'd just have stuff like the browser and chat windows. Though I might
even put the browser in the VM eventually.