macbook: add half-baked firefox support
Firefox is installed through a custom module. Home manager configuration
does work with it, but policy and extension installation does not. Thus
it doesn't really create a complete firefox setup.

Using homebrew would likely be easier but I didn't want to go to that
route. It's another software to maintain and historically it's not that
smooth. Plus I was not confident it would not pollute my PATH at some
point.

Current solution works good enough. I'm not really using the macbook for
any development. Even now I'm just fixing stuff out of boredom.