zsh: move to starship.rs
Circling back to a prompt manager. Previously I had some plugin for this
but ditched it as it was complex and slow.

At work I've had issues with various environment specific tools like
`kubectl` and `azure-cli`. It's often hard to know from the terminal
which environment I'm using. Starship had integrations for these tools
built in, which was enough for me.