desktop: increase nix module wankery
To be honest, I'm not sure if this is making things any better. Files
are now scattered around all over the place. Some in common, some in
host. While the copy-pasted solution had its warts, at least code
locality was high.

The flake configuration is nicer I admit that.

I'll take this as an experiment on getting comfortable with nix.

I'm still on the fence if home-manager is even worth it. I just use it
to install icons, Firefox profiles and maybe something else. I suspect I
could do most of this in nix.