doc: remove mentions of mac
I only ever boot the mac when I need to test one of the nixpkgs I
maintain. I haven't even been able to do that as the system is
constantly in broken state - I just never boot it and due to high
coupling changes in NixOS systems can break the macos system.