desktop: move to hyprland/wayland
Without any good reason I decided to have a non-working computer for a
couple of days and switch to wayland. I figured change would be nice.

This endeavour was not without pain. Hyprland documentation points to
git leading to a good few hours wasted on configuration that doesn't
exist in the released versions. Games would have like a 30-50% fps drop
for god knows why, but it seems to have been solved by not using
windowed mode.

Nixos side was not without pain as the nvidia tools are tied to the
xserver package, which apparently is not only for X11 but also for
wayland. Rebuilding the wayland based configuration can just get stuck.
Machine would go into a state where one cannot even execute reboot or
shutdown, the commands would just hang. This happens often when
switching from X11 to Wayland configuration.

Waybar CPU use shows garbage. It's around 0% most of the time but from
top I can see that it's not. Opening the calendar takes a couple of
seconds, how in the fuck does a text based calendar take a few seconds
to render?

Probably a dozen more shit I've just forgotten or hope to forget.