desktop: fix login to xterm instead of vtty
I think this fixes it. Login would randomly lead to a white background
shell, which I think is just the unstyled xterm. Exiting that shell
session would print the same kind of messages as exiting i3, hence why I
suspect xterm.

Why this happens completely randomly is a mystery. If I open another
vtty it may or may not use this shell again.